requirements.txt
pyspark>=3.0.0
lightgbm>=3.3.0
tensorflow>=2.10.0
grpcio>=1.48.0
grpcio-tools>=1.48.0
pandas>=1.3.0
scikit-learn>=1.0.0
